+++ ext/sgtelib/src/Surrogate_RBF.cpp
@@ -1,6 +1,6 @@
 /*-------------------------------------------------------------------------------------*/
 /*  sgtelib - A surrogate model library for derivative-free optimization               */
-/*  Version 2.0.2                                                                      */
+/*  Version 2.0.3                                                                      */
 /*                                                                                     */
 /*  Copyright (C) 2012-2017  Sebastien Le Digabel - Ecole Polytechnique, Montreal      */ 
 /*                           Bastien Talgorn - McGill University, Montreal             */
@@ -294,7 +294,7 @@ void SGTELIB::Surrogate_RBF::predict_private_objective ( const std::vector<SGTEL
   } // end for j
 
   // Loop on all pxx points 
-  for (int i=0 ; i<pxx ; i++){
+  for (int i=0 ; i<static_cast<int>(pxx) ; i++){
     // XXd[i] is of dimension nbd * _n
     ZZsurr_around->set_row( ( compute_design_matrix(*(XXd[i]), false) * _Alpha_obj ).transpose() , i );
   } // end for i
